人工智慧概論/AI_Final_Project/requirements.txt
numpy
gym
numpy
pytorch==1.12.1
torchvision==0.13.1 
torchaudio==0.12.1
transformers==4.5.1
wandb==0.9.1
tensorboard
pyprind
tensorflow
gin-config
gym==0.21.0
scikit-optimize
tqdm
blosc
git+https://github.com/google/dopamine.git
gym-anytrading